@@ -1,7 +1,10 @@
 {
-  "name": "new",
-  "version": " v1.22.17",
+  "name": "relife-resources",
+  "version": "1.1.0",
   "private": true,
+  "engines": {
+    "node": ">=20.0.0 <21.0.0"
+  },
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
@@ -9,27 +12,24 @@
     "@material-ui/icons": "^4.11.2",
     "@mui/material": "^5.15.11",
     "@mui/system": "^5.15.11",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.4",
-    "components": "^0.1.0",
-    "firebase": "^9.22.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "@reduxjs/toolkit": "^2.2.1",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.6.7",
+    "firebase": "^10.8.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.2",
+    "react-scripts": "5.0.1",
     "react-stripe-checkout": "^2.6.3",
     "redux-persist": "^6.0.0",
-    "styled": "^1.0.0",
-    "styled-components": "^6.0.0-rc.3",
+    "styled-components": "^6.1.8",
     "vercel": "^33.5.3",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
-    "dev":"react-scripts start ",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
