@@ -7,11 +7,8 @@ import Cart from "./pages/Cart";
 import NewProduct from "./pages/newProduct/NewProduct";
 import Success from "./pages/Success";
 import ProductAdded from "./pages/ProductAdded";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-} from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
 //  import { useSelector } from "react-redux";
 import Order from "./pages/Order";
 import Logout from "./pages/Logout";
@@ -20,56 +17,22 @@ import Logout from "./pages/Logout";
 const App = () => {
   //  const user=useSelector((state)=>state.user.currentUser);
   return (
-    
     <Router>
-      <Switch>
-      <Route exact path="/">
-           <Home/>
-         </Route>
-          <Route path="/login">
-            <Login/>
-         </Route>
-          <Route path="/home">
-            <Home/>
-         </Route>
-         
-         <Route path="/register">
-         <Register/>
-         </Route>
-      
-         <Route path="/products/:category">
-           <ProductList/>
-         </Route>
-         
-       <Route path="/newproduct">
-            <NewProduct />
-          </Route>
-         <Route path="/product/:id">
-           <Product/>
-         </Route>
-         
-      
-         <Route path="/cart">
-           <Cart/>
-         </Route>
-         <Route path="/success">
-          <Success/>
-         </Route>
-         <Route path="/logout">
-          <Logout/>
-         </Route>
-         <Route path="/productAdded">
-          <ProductAdded/>
-         </Route>
-         <Route path="/orderDetails">
-          <Order/>
-         </Route>
-      
-         
-     
-         
-      </Switch>
-    </Router>
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/home" element={<Home />} />
+      <Route path="/register" element={<Register />} />
+      <Route path="/products/:category" element={<ProductList />} />
+      <Route path="/newproduct" element={<NewProduct />} />
+      <Route path="/product/:id" element={<Product />} />
+      <Route path="/cart" element={<Cart />} />
+      <Route path="/success" element={<Success />} />
+      <Route path="/logout" element={<Logout />} />
+      <Route path="/productAdded" element={<ProductAdded />} />
+      <Route path="/orderDetails" element={<Order />} />
+    </Routes>
+  </Router>
   );
 };
 
