@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom/cjs/react-router-dom.min"
+import { Link } from "react-router-dom"
 import styled from "styled-components"
 
 const Container=styled.div`
