@@ -1,72 +1,69 @@
+import { createSlice } from "@reduxjs/toolkit";
 
-import {createSlice}from "@reduxjs/toolkit";
-
-const userSlice=createSlice({
-    name:"user",
-    initialState:{
-        currentUser:null,
-        isFetching:false,
-        error:false,
-        
+const userSlice = createSlice({
+  name: "user",
+  initialState: {
+    currentUser: null,
+    isFetching: false,
+    error: false,
+  },
+  reducers: {
+    loginStart: (state) => {
+      state.isFetching = true;
     },
-    reducers:{
-        loginStart:(state)=>{
-            state.isFetching=true
-            state.error=false
-            
-        },
-        loginSuccess:(state,action)=>{
-            state.isFetching=false;
-            state.currentUser=action.payload;
-            state.error=false
-        },
-        loginFailure:(state)=>{
-            state.isFetching=false;
-            state.error=true;
-        },
-        logout:(state)=>{
-            state.currentUser=null;
-            state.error=false;
-            
+    loginSuccess: (state, action) => {
+      state.isFetching = false;
+      state.currentUser = action.payload;
+    },
+    loginFailure: (state) => {
+      state.isFetching = false;
+      state.error = true;
+    },
+    logout: (state) => {
+      state.currentUser = null;
+    },
+    //ADD
+    addUserStart: (state) => {
+      state.isFetching = true;
+      state.error = false;
+    },
+    addUserSuccess: (state, action) => {
+      state.isFetching = false;
+      state.currentUser = action.payload;
+      state.error = false;
+      state.products.push(action.payload);
+      state.isAdmin = false;
+    },
+    addUserFailure: (state) => {
+      state.isFetching = false;
+      state.error = true;
+    },
+    //ADD ORDER DETAILS
+    addOrderStart: (state) => {
+      state.isFetching = true;
+      state.error = false;
+    },
+    addOrderSuccess: (state, action) => {
+      state.isFetching = false;
+      state.orders.push(action.payload);
+    },
+    addOrderFailure: (state) => {
+      state.isFetching = false;
+      state.error = true;
+    },
+  },
+});
 
-        },
-        //ADD
-        addUserStart: (state) => {
-            state.isFetching = true;
-            state.error = false;
-        },
-        addUserSuccess: (state,action) => {
-            state.isFetching = false;
-            state.currentUser=action.payload;
-            state.error=false;
-            state.products.push(action.payload);
-            state.isAdmin=false;
-        },
-        addUserFailure: (state) => {
-            state.isFetching = false;
-            state.error = true;
-        },
-        //ADD ORDER DETAILS
-        addOrderStart: (state) => {
-            state.isFetching = true;
-            state.error = false;
-        },
-        addOrderSuccess: (state, action) => {
-            
-            
-            state.isFetching = false;
-            state.orders.push(action.payload);
-        },
-        addOrderFailure: (state) => {
-            state.isFetching = false;
-            state.error = true;
-        },
-        
-        
-        },
-    });
-export const {loginStart,loginSuccess,loginFailure,
-                logout,
-                addUserFailure,addUserStart,addUserSuccess,addOrderFailure,addOrderStart,addOrderSuccess,
-            }=userSlice.actions;
+export const {
+  loginStart,
+  loginSuccess,
+  loginFailure,
+  logout,
+  addUserFailure,
+  addUserStart,
+  addUserSuccess,
+  addOrderFailure,
+  addOrderStart,
+  addOrderSuccess,
+} = userSlice.actions;
 export default userSlice.reducer;
